Add files via upload
This code is for a novel Arduino based lysimiter for agricultural research. The lysimiter is based on a cheap Chinese high precision balance with serial communication and interfaced with an Arduino Nano clone using a max 232 TTL converter. The serial data parsing strategy filters incomplete packets and other communication errors in real time.